+++ packages/wordpress-kit/src/lib/TailwindcssPlugin.ts
@@ -5,6 +5,7 @@ import {
   ImageComponent,
   PullQuoteComponent,
   TableComponent,
+  GalleryComponent,
 } from './tailwindcssPlugin/components';
 import { mergeToConfig } from './tailwindcssPlugin/Config';
 import { ColorUtilities, FontsUtilities } from './tailwindcssPlugin/utilities';
@@ -30,6 +31,10 @@ export default plugin(function ({ addUtilities, theme, addComponents }) {
     alignFull: { minWidth: theme('screen.xl', '1280px') },
   });
 
+  const gallery = GalleryComponent({
+    alignFull: { minWidth: theme('screen.xl', '1280px') },
+  });
+
   addUtilities([
     color.getBackgroundUtilities(),
     color.getBorderColorUtilities(),
@@ -41,5 +46,5 @@ export default plugin(function ({ addUtilities, theme, addComponents }) {
     quoteUtilities,
   ]);
 
-  addComponents([table, image, pullQuote]);
+  addComponents([table, image, pullQuote, gallery]);
 }, mergeToConfig);