+++ starters/next-drupal-starter/__tests__/getPaths.test.js
@@ -1,5 +1,5 @@
 import { getPaths } from "../lib/getPaths";
-import { globalDrupalStateStores } from "../lib/drupalStateContext";
+import { globalDrupalStateStores } from "../lib/stores";
 
 import umamiPaths from "./data/umamiPaths.json";
 import defaultProfilePaths from "./data/defaultProfilePaths.json";